+++ iFlex/Workout.swift
@@ -13,7 +13,7 @@ struct Workout {
     var time: TimeInterval      // specified in seconds -- is there a better data type?
     var exercises: [(Exercise, Int)]    // array of pairs: (exercise, number of reps)
     var iterations: Int
-    var difficulty: Float
+    var difficulty: Int         // value from 1-3: 1 = Easy, 2 = Medium, 3 = Hard
     var tags: [String]
     //var author: String
 }
